fix: show connection url instead of connect to server prompt for admin single-user catalog entry creation
