+++ ui/user/src/routes/mcp-catalog/EntriesView.svelte
@@ -1,13 +1,10 @@
 <script lang="ts">
 	import { tooltip } from '$lib/actions/tooltip.svelte';
-	import CopyButton from '$lib/components/CopyButton.svelte';
 	import DotDotDot from '$lib/components/DotDotDot.svelte';
-	import ResponsiveDialog from '$lib/components/ResponsiveDialog.svelte';
 	import ConnectToServer from '$lib/components/mcp/ConnectToServer.svelte';
 	import McpConfirmDelete from '$lib/components/mcp/McpConfirmDelete.svelte';
 	import McpMultiDeleteBlockedDialog from '$lib/components/mcp/McpMultiDeleteBlockedDialog.svelte';
 	import StaticOAuthConfigureModal from '$lib/components/mcp/StaticOAuthConfigureModal.svelte';
-	import IconButton from '$lib/components/primitives/IconButton.svelte';
 	import Table, { type InitSort, type InitSortFn } from '$lib/components/table/Table.svelte';
 	import {
 		AdminService,
@@ -32,23 +29,20 @@
 	import { mcpServersAndEntries, profile } from '$lib/stores';
 	import { formatTimeAgo } from '$lib/time';
 	import { openUrl } from '$lib/utils';
+	import McpConnectUrlDialog from './McpConnectUrlDialog.svelte';
 	import {
 		Captions,
 		CircleFadingArrowUp,
 		Ellipsis,
 		GitBranch,
-		Info,
 		Link2Icon,
-		Plus,
 		Server,
 		Settings,
-		StepForward,
 		Trash2,
 		TriangleAlert,
 		UsersIcon
 	} from 'lucide-svelte';
 	import type { Snippet } from 'svelte';
-	import { twMerge } from 'tailwind-merge';
 
 	type Item = ReturnType<typeof convertEntriesToTableData>[number];
 
@@ -94,15 +88,7 @@
 	let deleteConflictError = $state<MCPCompositeDeletionDependencyError | undefined>();
 
 	let connectToServerDialog = $state<ReturnType<typeof ConnectToServer>>();
-	let displayConnectUrl = $state<{ url: string; entry?: MCPCatalogEntry }>();
-	let connectUrlDialog = $state<ReturnType<typeof ResponsiveDialog>>();
-	let copyButton = $state<ReturnType<typeof CopyButton>>();
-
-	let selectServerForCatalogEntry = $state<{
-		entry: MCPCatalogEntry;
-		servers: MCPCatalogServer[];
-	}>();
-	let selectServerDialog = $state<ReturnType<typeof ResponsiveDialog>>();
+	let connectUrlDialog = $state<ReturnType<typeof McpConnectUrlDialog>>();
 
 	let oauthConfigModal = $state<ReturnType<typeof StaticOAuthConfigureModal>>();
 	let oauthConfigEntry = $state<MCPCatalogEntry>();
@@ -174,25 +160,12 @@
 		}
 
 		if (server?.connectURL) {
-			displayConnectUrl = { url: server.connectURL, entry };
-			connectUrlDialog?.open();
+			connectUrlDialog?.open(entry, server.connectURL);
 		}
-	}
-
-	function handleShowSelectServerDialog(entry: MCPCatalogEntry, servers: MCPCatalogServer[]) {
-		selectServerForCatalogEntry = {
-			entry,
-			servers
-		};
-		selectServerDialog?.open();
 	}
 
 	function getMultiUserCatalogEntryServers(entry: MCPCatalogEntry) {
 		return mcpServersAndEntries.current.servers.filter((s) => s.catalogEntryID === entry.id);
-	}
-
-	function isConfigurableSingleUserCatalogEntry(entry?: MCPCatalogEntry) {
-		return entry && !isMultiUserCatalogEntry(entry) && hasEditableConfiguration(entry);
 	}
 
 	async function handleConfigureOAuth(entry: MCPCatalogEntry) {
@@ -423,29 +396,7 @@
 										class="menu-button"
 										onclick={(e) => {
 											e.stopPropagation();
-											if (isMultiUserCatalogEntry(catalogEntry)) {
-												// multi user catalog requires configuration of at least one server to obtain connect URL
-												const matchingServers = getMultiUserCatalogEntryServers(catalogEntry);
-												if (matchingServers.length > 1) {
-													handleShowSelectServerDialog(catalogEntry, matchingServers);
-												} else if (matchingServers[0]?.connectURL) {
-													displayConnectUrl = {
-														url: matchingServers[0].connectURL,
-														entry: catalogEntry
-													};
-													connectUrlDialog?.open();
-												} else {
-													connectToServerDialog?.open({ entry: catalogEntry });
-												}
-											} else {
-												// single user catalog entry contains baseline connect URL
-												displayConnectUrl = {
-													url: catalogEntry?.connectURL ?? '',
-													entry: catalogEntry
-												};
-												connectUrlDialog?.open();
-											}
-
+											connectUrlDialog?.open(catalogEntry);
 											toggle(false);
 										}}
 									>
@@ -499,69 +450,12 @@
 	entityPlural="entries"
 />
 
-<ResponsiveDialog
+<McpConnectUrlDialog
 	bind:this={connectUrlDialog}
-	animate="slide"
-	title="Connection URL"
-	class="max-w-[95vw] md:max-w-2xl"
-	classes={{ content: 'p-0', header: 'p-4 pb-0' }}
-	disableMobileStyles
->
-	<div class={twMerge('px-4', !isMultiUserCatalogEntry(displayConnectUrl?.entry) && 'pb-4')}>
-		<div class="rounded-field bg-base-200 border-none input w-full pr-0">
-			<div class="label px-2.5 flex items-center gap-2 text-xs text-base-content/75 shrink-0 mr-1">
-				<Link2Icon class="size-4" />
-			</div>
-			<input
-				onmousedown={() => copyButton?.copy()}
-				type="text"
-				value={displayConnectUrl?.url ?? ''}
-				class="w-full text-xs"
-				readonly
-			/>
-			<div class="mr-2">
-				<CopyButton
-					bind:this={copyButton}
-					classes={{
-						button:
-							'shrink-0 text-xs flex gap-1 :not([disabled]):hover:text-base-content :not([disabled]):text-base-content/65 disabled:cursor-not-allowed'
-					}}
-					text={displayConnectUrl?.url ?? ''}
-					showTextLeft
-				/>
-			</div>
-		</div>
-	</div>
-	{#if isMultiUserCatalogEntry(displayConnectUrl?.entry)}
-		<div class="mt-4 p-4 border-t border-base-300 dark:border-base-400">
-			<p
-				class="text-muted-content flex items-center justify-end gap-2 text-sm font-light md:px-0 px-4"
-			>
-				Need to set up a different instance?
-				<button
-					class="btn btn-sm btn-primary text-xs"
-					onclick={() => {
-						if (displayConnectUrl?.entry) {
-							connectUrlDialog?.close();
-							connectToServerDialog?.open({ entry: displayConnectUrl.entry });
-						}
-					}}
-				>
-					<Plus class="size-3" />
-					Launch New Server
-				</button>
-			</p>
-		</div>
-	{:else if isConfigurableSingleUserCatalogEntry(displayConnectUrl?.entry)}
-		<div class="notification-info m-4 mt-0">
-			<p class="flex items-center gap-2 text-xs">
-				<Info class="size-4" />
-				This server requires end-user configuration. Users will need to configure their instance of it
-				here before they can connect to it.
-			</p>
-		</div>
-	{/if}
-</ResponsiveDialog>
+	onLaunchCatalogEntry={(entry) => {
+		connectToServerDialog?.open({ entry });
+	}}
+/>
 
 <McpConfirmDelete
 	names={[getMCPDisplayName(deletingServer)]}
@@ -650,45 +544,3 @@
 	onSave={handleSaveOAuth}
 	onDelete={handleDeleteOAuth}
 />
-
-<ResponsiveDialog
-	class="bg-base-200 dark:bg-base-100"
-	bind:this={selectServerDialog}
-	title="Select Your Server"
->
-	<Table
-		data={selectServerForCatalogEntry?.servers || []}
-		fields={['name', 'created']}
-		onClickRow={async (d) => {
-			selectServerDialog?.close();
-			if (!d.connectURL) return;
-			displayConnectUrl = { url: d.connectURL, entry: selectServerForCatalogEntry?.entry };
-			connectUrlDialog?.open();
-		}}
-		disablePortal
-	>
-		{#snippet onRenderColumn(property, d)}
-			{#if property === 'name'}
-				<div class="flex shrink-0 items-center gap-2">
-					<div class="icon">
-						{#if d.manifest.icon}
-							<img src={d.manifest.icon} alt={d.manifest.name} class="size-6" />
-						{:else}
-							<Server class="size-6" />
-						{/if}
-					</div>
-					<p class="flex items-center gap-2">
-						{getMCPDisplayName(d)}
-					</p>
-				</div>
-			{:else if property === 'created'}
-				{formatTimeAgo(d.created).relativeTime}
-			{/if}
-		{/snippet}
-		{#snippet actions()}
-			<IconButton class="hover:dark:bg-base-100/50">
-				<StepForward class="size-4" />
-			</IconButton>
-		{/snippet}
-	</Table>
-</ResponsiveDialog>
